Created controller package
This package is a replacement for the behavior package. It is intended to
implement SMACS (Stack of Modular Architecture Control States) system. This
system is a behavior stack unlike a behavior queue that is implemented now.

It will consist of handlers, controllers, and behaviors. The MainLoop.cpp
will be the initializer of all the handlers, controllers, and behaviors. It
will also serve as a ticker (it will call update in the behaviors).
This is the big important class basically.