Implemented simple walk.
Created a drive controller. The controller can drive the robots to (x, y)
location. Uses the PID that UNM gave. Also made it so that the robots correct
their driving direction while in drive mode. This seems to make the robots
look like they try to go in straight lines.